+++ e2e/contracts/testdappv2/bindings.go
