+++ e2e/e2etests/test_deploy_contract.go
@@ -2,13 +2,13 @@ package e2etests
 
 import (
 	"fmt"
+	"github.com/zeta-chain/node/e2e/contracts/testdappv2"
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
-	"github.com/zeta-chain/node/pkg/contracts/testdappv2"
 )
 
 // deployFunc is a function that deploys a contract
