+++ e2e/e2etests/test_zevm_to_evm_call_through_contract.go
@@ -1,6 +1,7 @@
 package e2etests
 
 import (
+	"github.com/zeta-chain/node/e2e/contracts/gatewayzevmcaller"
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
@@ -8,7 +9,6 @@ import (
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
-	gatewayzevmcaller "github.com/zeta-chain/node/pkg/contracts/gatewayzevmcaller"
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 )
