+++ e2e/runner/setup_evm.go
@@ -1,6 +1,8 @@
 package runner
 
 import (
+	"github.com/zeta-chain/node/e2e/contracts/erc1967proxy"
+	"github.com/zeta-chain/node/e2e/contracts/testdappv2"
 	"math/big"
 	"time"
 
@@ -12,8 +14,6 @@ import (
 
 	"github.com/zeta-chain/node/e2e/contracts/erc20"
 	"github.com/zeta-chain/node/e2e/utils"
-	"github.com/zeta-chain/node/pkg/contracts/erc1967proxy"
-	"github.com/zeta-chain/node/pkg/contracts/testdappv2"
 )
 
 // SetupEVM setup contracts on EVM with v2 contracts
