+++ e2e/runner/testdapp.go
@@ -1,13 +1,12 @@
 package runner
 
 import (
+	"github.com/zeta-chain/node/e2e/contracts/testdappv2"
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
-
-	"github.com/zeta-chain/node/pkg/contracts/testdappv2"
 )
 
 // AssertTestDAppZEVMCalled is a function that asserts the values of the test dapp on the ZEVM
