+++ precompiles/bank/method_test.go
@@ -1,6 +1,7 @@
 package bank
 
 import (
+	"github.com/zeta-chain/node/e2e/contracts/erc1967proxy"
 	"math/big"
 	"testing"
 
@@ -18,7 +19,6 @@ import (
 	evmkeeper "github.com/zeta-chain/ethermint/x/evm/keeper"
 	"github.com/zeta-chain/ethermint/x/evm/statedb"
 	"github.com/zeta-chain/node/pkg/chains"
-	"github.com/zeta-chain/node/pkg/contracts/erc1967proxy"
 	"github.com/zeta-chain/node/pkg/ptr"
 	precompiletypes "github.com/zeta-chain/node/precompiles/types"
 	"github.com/zeta-chain/node/testutil/keeper"
