+++ precompiles/staking/staking_test.go
@@ -2,6 +2,7 @@ package staking
 
 import (
 	"encoding/json"
+	"github.com/zeta-chain/node/e2e/contracts/erc1967proxy"
 	"testing"
 
 	"math/big"
@@ -28,7 +29,6 @@ import (
 	"github.com/zeta-chain/ethermint/x/evm/statedb"
 	"github.com/zeta-chain/node/cmd/zetacored/config"
 	"github.com/zeta-chain/node/pkg/chains"
-	"github.com/zeta-chain/node/pkg/contracts/erc1967proxy"
 	"github.com/zeta-chain/node/pkg/ptr"
 	"github.com/zeta-chain/node/precompiles/prototype"
 	"github.com/zeta-chain/node/testutil/keeper"
